tools/image_review/requirements.txt
flask>=2.0
pillow>=9.0
pymongo>=4.6.0
numpy>=1.24.0
